+++ CloudServer/src/Core/Connection/TCPListener.h
@@ -32,27 +32,27 @@ public:
 
 protected:
 	// Handler for a new connections
-	virtual void onClientConnected(uint32_t clientSocket);
+	virtual void onClientConnected(uint64_t clientSocket);
 
 	// Handler for disconnecting 
-	virtual void onClientDisconnected(uint32_t clientSocket);
+	virtual void onClientDisconnected(uint64_t clientSocket);
 
 	// Handler pentru un mesaj primit de la client
-	virtual void onMessageReceived(uint32_t clientSocket, std::string& msg, int length);
+	virtual void onMessageReceived(uint64_t clientSocket, std::string& msg, int length);
 
 	// Sending message to a client
-	void sendToClient(uint32_t clientSocket, const char* msg, size_t length);
+	void sendToClient(uint64_t clientSocket, const char* msg, size_t length);
 
 	//Sending message to a client
-	void sendToClient(uint32_t clientSocket, std::string& msg, size_t length);
+	void sendToClient(uint64_t clientSocket, std::string& msg, size_t length);
 
 	//Sending message to all clients
-	void broadcastToClients(uint32_t sendingClient, const char* msg, int length);
+	void broadcastToClients(uint64_t sendingClient, const char* msg, int length);
 private:
 
 	const char* s_ipAddress;					// Adresa IP a serverului 
 	int				s_port;						// Nr. portului
-	uint32_t		s_socket;					// Internal FD for the listening socket
+	uint64_t		s_socket;					// Internal FD for the listening socket
 	fd_set			master;						// Master file descriptor set
 
 protected:
